refactor: move `IndexPackage` and `RegistryDependency` to `cargo-util-schemas` for better modularity
