+++ crates/cargo-util-schemas/Cargo.toml
@@ -25,6 +25,7 @@ workspace = true
 
 [dev-dependencies]
 snapbox.workspace = true
+serde_json.workspace = true
 
 [features]
 unstable-schema = ["dep:schemars", "dep:serde_json"]
