+++ crates/cargo-util-schemas/src/lib.rs
@@ -9,6 +9,7 @@
 //! > ecosystem. This crate follows semver compatibility for its APIs.
 
 pub mod core;
+pub mod index;
 pub mod manifest;
 pub mod messages;
 #[cfg(feature = "unstable-schema")]
